% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{CNV.alteration}
\alias{CNV.alteration}
\title{CNV.alteration}
\usage{
CNV.alteration(
  x,
  threshold = 0.1,
  states = c("loss", "balanced", "gain"),
  deepthresh = Inf
)
}
\arguments{
\item{x}{a vector of segment CN values}

\item{threshold}{a threshold magnitude to call an alteration (0.1)}

\item{states}{a vector of state names (default "loss","balanced","gain")}

\item{deepthresh}{magnitude for calling hiamp/deeploss (Inf)}
}
\value{
a vector of alteration calls using the above
}
\description{
helper function for converting +/- threshold to Alteration calls
}
\examples{
Seg.CN <- c(A=-0.2, B=-0.1, C=-0.05, D=0, E=0.05, F=0.1, G=0.2)
CNV.alteration(Seg.CN, threshold=.05)
CNV.alteration(Seg.CN, threshold=0.1, states=c("del", "neu", "amp"))
as.numeric(CNV.alteration(Seg.CN, threshold=0.2, states=c("-1", "0", "+1")))
CNV.alteration(Seg.CN, threshold=1)

}
