% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{CNV.analysis-class}
\alias{CNV.analysis-class}
\alias{show,CNV.analysis-method}
\alias{coef,CNV.analysis-method}
\alias{[,CNV.analysis,ANY,ANY,ANY-method}
\title{CNV.analysis class}
\usage{
\S4method{show}{CNV.analysis}(object)

\S4method{coef}{CNV.analysis}(object)

\S4method{[}{CNV.analysis,ANY,ANY,ANY}(x, i)
}
\arguments{
\item{object}{\code{CNV.analysis} object}

\item{x}{\code{CNV.analysis} object (defined by \code{Extract} generic).}

\item{i}{index. \code{logical}, \code{numeric} or \code{character}.}
}
\value{
\code{CNV.analysis} class.
}
\description{
CNV analysis data of a single sample is stored in this class
}
\details{
Use \code{CNV.fit} to create. Modified by \code{CNV.bin}, \code{CNV.detail} and \code{CNV.segment}.
}
\examples{
# general information
x
show(x)

# coefficients of linear regression
coef(x)

# show or replace sample name
names(x)
names(x) <- c('Sample.1', 'Sample.2')

# subset samples

x[1]
# output plots
CNV.genomeplot(x)
CNV.genomeplot(x, chr = 'chr6')
#CNV.detailplot(x, name = 'MYCN')
#CNV.detailplot_wrap(x)
CNV.write(x, what = 'segments')
}
\author{
Volker Hovestadt, Bjarne Daenekas \email{conumee@hovestadt.bio}
}
