% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{CNV.combine}
\alias{CNV.combine}
\alias{CNV.combine,CNV.analysis,CNV.analysis-method}
\title{CNV.combine}
\usage{
CNV.combine(query.1, query.2, ...)

\S4method{CNV.combine}{CNV.analysis,CNV.analysis}(query.1, query.2)
}
\arguments{
\item{query.1}{\code{CNV.analysis} object 1.}

\item{query.2}{\code{CNV.analysis} object 2.}

\item{...}{Additional parameters (\code{CNV.combine} generic, currently not used).}
}
\value{
\code{CNV.analysis} object.
}
\description{
Combine two \code{CNV.analysis} objects. Please make sure that both objects were generated with the same reference samples and the same annotation object in \code{CNV.fit()}.
}
\details{
This function combines two \code{CNV.analysis} objects. This is necessary if summaryplots or heatmaps should be generated for a cohort that comprises methylation profiles from multiple array types.

x.1 <- CNV.fit(query.1, ref, anno)
x.2 <- CNV.fit(query.2, ref, anno)

x.combined <- CNV.combine(x.1, x.2)
}
\author{
Bjarne Daenekas \email{conumee@hovestadt.bio}
}
