% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{CNV.create_bins}
\alias{CNV.create_bins}
\title{CNV.create_bins}
\usage{
CNV.create_bins(genome.anno, bin_minsize = 50000, genome.gap, genome.exclude)
}
\arguments{
\item{genome.anno}{foo}

\item{bin_minsize}{foo}

\item{genome.gap}{foo}

\item{genome.exclude}{foo}
}
\value{
\code{GRanges} object.
}
\description{
Split genome into bins of defined size.
}
