% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{CNV.define_detail}
\alias{CNV.define_detail}
\title{CNV.define_detail}
\usage{
CNV.define_detail(array_type = "450k", gene = "predefined")
}
\arguments{
\item{array_type}{character. One of \code{450k}, \code{EPIC}, \code{EPICv2} or \code{mouse}. Defaults to \code{450k}.}

\item{gene}{character vector. Specify the genes by using their \code{SYMBOL} (case sensitive). Default to \code{predefined} which loads a predefined set of onco- and tumor suppressor genes.}

\item{...}{Additional parameters (\code{CNV.load} generic, currently not used).}
}
\value{
\code{GRanges} object
}
\description{
Create a \code{GRanges} object for detail regions. Downstream plotting functions will create individual plots for each detail region.
}
\examples{
#create a GRanges object of detail regions
detail_regions <- CNV.define_detail(array_type = "450k", symbol = c("CDK6", "PTEN", "MYCN"))
detail_regions
}
\author{
Bjarne Daenekas \email{conumee@hovestadt.bio}
}
