% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{CNV.detailplot}
\alias{CNV.detailplot}
\alias{CNV.detailplot,CNV.analysis-method}
\title{CNV.detailplot}
\usage{
CNV.detailplot(object, ...)

\S4method{CNV.detailplot}{CNV.analysis}(
  object,
  name,
  yaxt = "l",
  ylim = c(-1.25, 1.25),
  set_par = TRUE,
  output = NULL,
  columns = NULL,
  main = NULL,
  directory = getwd(),
  width = 12,
  height = 8,
  res = 720,
  cols = c("darkblue", "darkblue", "lightgrey", "#F16729", "#F16729")
)
}
\arguments{
\item{object}{\code{CNV.analysis} object.}

\item{...}{Additional parameters (\code{CNV.detailplot} generic, currently not used).}

\item{name}{character. Name of detail region to plot.}

\item{yaxt}{character. Include y-axis? \code{'l'}: left, \code{'r'}: right, \code{'n'}: no. Defaults to \code{'l'}.}

\item{ylim}{numeric vector. The y limits of the plot. Defaults to \code{c(-1.25, 1.25)}.}

\item{set_par}{logical. Use recommended graphical parameters for \code{oma} and \code{mar}? Defaults to \code{TRUE}. Original parameters are restored afterwards.}

\item{output}{character. Choose between \code{pdf} and \code{png}. Defaults to \code{NULL}}

\item{columns}{numeric. Needed for \code{detailplot_wrap}. Defaults to \code{NULL}. Do not manipulate.}

\item{directory}{character. Export directory for saving the files}

\item{width}{numeric. Width in inches of the saved files. Defaults to \code{12}.}

\item{height}{numeric. Height in inches of the saved files. Defaults to \code{8}}

\item{res}{numeric. Resolution of the saved .png files. Defaults to \code{720}}

\item{cols}{character vector. Colors to use for plotting intensity levels of bins. Centered around 0. Defaults to \code{c('red', 'red', 'lightgrey', 'green', 'green')}.}
}
\value{
\code{NULL}.
}
\description{
Create CNV plot for detail region. If the \code{CNV.analysis} object holds the information for multiple samples, the plots get either loaded individually in the graphical output or directly saved as .pdf or .png files.
}
\details{
This method provides the functionality for generating detail regions CNV plots. Probes are shown as dots, bins are shown as lines. See parameters for more information.
}
\examples{
# prepare
library(minfiData)
data(MsetEx)
d <- CNV.load(MsetEx)
data(detail_regions)
anno <- CNV.create_anno(detail_regions = detail_regions)

# create/modify object
x <- CNV.segment(CNV.detail(CNV.bin(CNV.fit(query = d['GroupB_1'],
    ref = d[c('GroupA_1', 'GroupA_2', 'GroupA_3')], anno))))

# output plots
CNV.genomeplot(x)
CNV.genomeplot(x, chr = 'chr6')
CNV.detailplot(x, name = 'PTEN', output = "pdf", directory = dir)
CNV.detailplot_wrap(x)
CNV.summaryplot(x)

# output text files
CNV.write(x, what = 'segments')
CNV.write(x, what = 'detail')
CNV.write(x, what = 'bins')
CNV.write(x, what = 'probes')
}
\author{
Volker Hovestadt, Bjarne Daenekas \email{conumee@hovestadt.bio}
}
