% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{CNV.detailplot_wrap}
\alias{CNV.detailplot_wrap}
\alias{CNV.detailplot_wrap,CNV.analysis-method}
\title{CNV.detailplot_wrap}
\usage{
CNV.detailplot_wrap(object, ...)

\S4method{CNV.detailplot_wrap}{CNV.analysis}(
  object,
  set_par = TRUE,
  main = NULL,
  header = NULL,
  output = NULL,
  directory = getwd(),
  width = 12,
  height = 8,
  res = 720,
  ...
)
}
\arguments{
\item{object}{\code{CNV.analysis} object.}

\item{...}{Additional paramters supplied to \code{CNV.detailplot}.}

\item{set_par}{logical. Use recommended graphical parameters for \code{oma} and \code{mar}? Defaults to \code{TRUE}. Original parameters are restored afterwards.
#' @param directory character. Export directory for saving the files}

\item{main}{character. Used for \code{CNV.detailplot}. do not manipulate}

\item{header}{character vector. Title of the plot(s). Defaults to sample names. Please provide a vector of the same length than the number of samples.}

\item{output}{character. Choose between \code{pdf} and \code{png}. Defaults to \code{NULL}}

\item{width}{numeric. Width in inches of the saved files. Defaults to \code{12}.}

\item{height}{numeric. Height in inches of the saved files. Defaults to \code{8}}

\item{res}{numeric. Resolution of the saved .png files. Defaults to \code{720}}
}
\value{
\code{NULL}.
}
\description{
Create CNV plots for all detail regions. If the \code{CNV.analysis} object holds the information for multiple samples, the plots get either loaded individually in the graphical output or directly saved as .pdf or .png files.
}
\details{
This method is a wrapper of the \code{CNV.detailplot} method to plot all detail regions.
}
\examples{
# prepare
library(minfiData)
data(MsetEx)
d <- CNV.load(MsetEx)
data(detail_regions)
anno <- CNV.create_anno(detail_regions = detail_regions)

# create/modify object
x <- CNV.segment(CNV.detail(CNV.bin(CNV.fit(query = d['GroupB_1'],
    ref = d[c('GroupA_1', 'GroupA_2', 'GroupA_3')], anno))))

# output plots
CNV.genomeplot(x)
CNV.genomeplot(x, chr = 'chr6')
CNV.detailplot(x, name = 'PTEN')
CNV.detailplot_wrap(x)
CNV.summaryplot(x)

# output text files
CNV.write(x, what = 'segments')
CNV.write(x, what = 'detail')
CNV.write(x, what = 'bins')
CNV.write(x, what = 'probes')
}
\author{
Volker Hovestadt, Bjarne Daenekas \email{conumee@hovestadt.bio}
}
