% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{CNV.df}
\alias{CNV.df}
\alias{CNV.df,CNV.data-method}
\title{CNV.df}
\usage{
CNV.df(object, anno)

\S4method{CNV.df}{CNV.data}(object, anno)
}
\arguments{
\item{object}{\code{CNV.data} object.}

\item{anno}{\code{CNV.anno} object.}
}
\value{
\code{CNV.data} object.
}
\description{
Turn the probe intensities within a \code{CNV.data} object (list) into a dataframe.
}
\author{
Bjarne Daenekas \email{conumee@hovestadt.bio}
}
