% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{CNV.focal}
\alias{CNV.focal}
\alias{CNV.focal,CNV.analysis-method}
\title{CNV.focal}
\usage{
CNV.focal(object, ...)

\S4method{CNV.focal}{CNV.analysis}(
  object,
  sig_cgenes = FALSE,
  conf = 0.99,
  R = 100,
  blockLength = 5e+05,
  proportionLength = TRUE,
  ...
)
}
\arguments{
\item{object}{\code{CNV.analysis} object.}

\item{...}{Additional parameters (\code{CNV.detailplot} generic, currently not used).}

\item{sig_cgenes}{logical. Should the genes from the Cancer Gene Census be assessed as well? (The high number of genes can lead to false positive results.) Default to \code{FALSE}.}

\item{conf}{numeric. Confidence level to determine the log2-threshold for focal alterations. Default to \code{0.99}.}

\item{R}{numeric. Parameter for the \code{bootRanges} function. The number of bootstrap samples to generate. Default to \code{100}.}

\item{blockLength}{numeric. Parameter for the \code{bootRanges} function. The length (in basepairs) of the blocks for Segmented Block Bootstrapping. Default to \code{500000}.}

\item{proportionLength}{logical. From the \code{nullranges} package: for the segmented block bootstrap, whether to use scaled block lengths, (scaling by the proportion of the segmentation state out of the total genome length)}
}
\value{
A \code{CNV.analysis} object with significantly altered bins and predefined focal regions of interest.
}
\description{
This function aims to detect focal CNVs that are defined as \code{detail_regions} in the annotation object.
}
\details{
This function should facilitate the detection of CNVs that affect single genes. Segmented Block Bootstrapping is applied to model a normal distribution of bins for each copy-number state. Two-sided confidence intervals are applied to calculate dynamic thresholds for amplifications and losses.
}
\examples{

x <- CNV.focal(x, conf = 0.99, R = 100, blockLength = 500000)
x@detail$del.bins  #bins that are part of deletions.
x@detail$amp.bins  #bins that are part of amplifications.
x@detail$del.detail.regions #deleted predefined regions
x@detail$amp.detail.regions #amplified predefined regions


}
\author{
Bjarne Daenekas, Volker Hovestadt \email{conumee@hovestadt.bio}
}
