% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{CNV.genomeplot}
\alias{CNV.genomeplot}
\alias{CNV.genomeplot,CNV.analysis-method}
\title{CNV.genomeplot}
\usage{
CNV.genomeplot(object, ...)

\S4method{CNV.genomeplot}{CNV.analysis}(
  object,
  chr = "all",
  chrX = TRUE,
  chrY = TRUE,
  centromere = TRUE,
  detail = TRUE,
  main = NULL,
  sig_cgenes = TRUE,
  nsig_cgenes = 5,
  output = NULL,
  directory = getwd(),
  ylim = c(-1.25, 1.25),
  bins_cex = "standardized",
  set_par = TRUE,
  width = 12,
  height = 6,
  res = 720,
  cols = c("darkblue", "darkblue", "lightgrey", "#F16729", "#F16729")
)
}
\arguments{
\item{object}{\code{CNV.analysis} object.}

\item{...}{Additional parameters (\code{CNV.detailplot} generic, currently not used).}

\item{chr}{character vector. Which chromosomes to plot. Defaults to \code{'all'}.}

\item{centromere}{logical. Show dashed lines at centromeres? Defaults to \code{TRUE}.}

\item{detail}{logical. If available, include labels of detail regions? Defaults to \code{TRUE}.}

\item{main}{character vector. Title of the plot(s). Defaults to sample names. Please provide a vector of the same length as the number of samples.}

\item{sig_cgenes}{logical. Should the significant genes be plotted that were identified with \code{CNV.focal}? Default to \code{TRUE}.}

\item{nsig_cgenes}{numeric. How many significant genes identified with \code{CNV.focal} should be plotted? Default to \code{5}.}

\item{output}{character. Choose between \code{pdf} and \code{png}. Defaults to \code{NULL}}

\item{directory}{character. Export directory for saving the files}

\item{ylim}{numeric vector. The y limits of the plot. Defaults to \code{c(-1.25, 1.25)}.}

\item{bins_cex}{character. The size of the individual bin dots is reversely proportional its variance of included probes' log2-ratios. Choose either \code{standardized} for fixed dot sizes (to make plots from different samples comparable) or \code{sample_level} (to scale the dot sizes for each sample individually). Default to \code{standardized}.}

\item{set_par}{logical. Use recommended graphical parameters for \code{oma} and \code{mar}? Defaults to \code{TRUE}. Original parameters are restored afterwards.}

\item{width}{numeric. Width in inches of the saved files. Defaults to \code{12}.}

\item{height}{numeric. Height in inches of the saved files. Defaults to \code{8}}

\item{res}{numeric. Resolution of the saved .png files. Defaults to \code{720}}

\item{cols}{character vector. Colors to use for plotting intensity levels of bins. Centered around 0. Defaults to \code{c('red', 'red', 'lightgrey', 'green', 'green')}.}
}
\value{
\code{NULL}.
}
\description{
Create CNV plot for the whole genome or chromosomes. If the \code{CNV.analysis} object holds the information for multiple samples, the plots get either loaded individually in the graphical output or directly saved as .pdf or .png files.
}
\details{
This method provides the functionality for generating CNV plots for the whole genome or defined chromosomes. Bins are shown as dots, segments are shown as lines. See parameters for more information.
}
\examples{
# prepare
library(minfiData)
data(MsetEx)
d <- CNV.load(MsetEx)
data(detail_regions)
anno <- CNV.create_anno(detail_regions = detail_regions)

# create/modify object
x <- CNV.segment(CNV.detail(CNV.bin(CNV.fit(query = d['GroupB_1'],
    ref = d[c('GroupA_1', 'GroupA_2', 'GroupA_3')], anno))))

# output plots
CNV.genomeplot(x)
CNV.genomeplot(x, chr = 'chr6')
CNV.genomeplot(x, output = "pdf", directory = dir)
CNV.detailplot(x, name = 'PTEN')
CNV.detailplot_wrap(x)
CNV.summaryplot(x)

# output text files
CNV.write(x, what = 'segments')
CNV.write(x, what = 'detail')
CNV.write(x, what = 'bins')
CNV.write(x, what = 'probes')
}
\author{
Volker Hovestadt, Bjarne Daenekas \email{conumee@hovestadt.bio}
}
