% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{CNV.heatmap}
\alias{CNV.heatmap}
\alias{CNV.heatmap,CNV.analysis-method}
\title{CNV.heatmap}
\usage{
CNV.heatmap(object, ...)

\S4method{CNV.heatmap}{CNV.analysis}(
  object,
  set_par = TRUE,
  main = NULL,
  hclust = TRUE,
  hclust_method = "average",
  dist_method = "euclidian",
  cexRow = 1/2,
  zlim = c(-0.5, 0.5),
  useRaster = TRUE,
  ...
)
}
\arguments{
\item{object}{\code{CNV.analysis} object.}

\item{...}{Additional parameters}

\item{set_par}{logical. Use recommended graphical parameters for \code{oma} and \code{mar}? Defaults to \code{TRUE}. Original parameters are restored afterwards.}

\item{main}{character. Specify the title of the plot. Defaults to \code{NULL}.}

\item{hclust}{logical. Should hierarchical clustering be performed? Default to \code{TRUE}.}

\item{hclust_method}{character. The agglomeration method to be used. This should be (an unambiguous abbreviation of) one of "ward.D", "ward.D2", "single", "complete", "average" (= UPGMA), "mcquitty" (= WPGMA), "median" (= WPGMC) or "centroid" (= UPGMC).}

\item{dist_method}{character. The distance measure to be used. This must be one of "euclidean", "maximum", "manhattan", "canberra", "binary" or "minkowski". Any unambiguous substring can be given.}

\item{cexRow}{numeric. Adjust the font size for row labeling. Default to 0.5.}

\item{zlim}{numeric. The minimum and maximum z values for which colors should be plotted. Default to \code{c(-0.5,0.5)}.}

\item{useRaster}{logical. Should a bitmap raster be used to create the plot instead of polygons? Default to \code{TRUE}.}
}
\description{
Create a heatmap to illustrate CNVs in a set of query samples. Colors correspond to the other plots.
}
\examples{
#' # prepare
library(minfiData)
data(MsetEx)
d <- CNV.load(MsetEx)
data(detail_regions)
anno <- CNV.create_anno(detail_regions = detail_regions)

# create/modify object
x <- CNV.segment(CNV.detail(CNV.bin(CNV.fit(query = d['GroupB_1'],
    ref = d[c('GroupA_1', 'GroupA_2', 'GroupA_3')], anno))))

# output plots
CNV.genomeplot(x)
CNV.genomeplot(x, chr = 'chr6')
CNV.detailplot(x, name = 'PTEN')
CNV.detailplot_wrap(x)
CNV.summaryplot(x)
CNV.heatmap(x)

# output text files
CNV.write(x, what = 'segments')
CNV.write(x, what = 'detail')
CNV.write(x, what = 'bins')
CNV.write(x, what = 'probes')
}
\author{
Bjarne Daenekas \email{conumee@hovestadt.bio}
}
