% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{CNV.import}
\alias{CNV.import}
\alias{CNV.import,character,data.frame-method}
\title{CNV.import}
\usage{
CNV.import(directory, sample_sheet, ...)

\S4method{CNV.import}{character,data.frame}(directory = getwd(), sample_sheet = NULL)
}
\arguments{
\item{directory}{Specify the folder that stores the .idat-Files.}

\item{sample_sheet}{dataframe. Provide a sample sheet with at least three columns: \code{Sample_Name}, \code{Sentrix_ID} and \code{Sentrix_Position}. The spelling of the colnames must be exactly as shown.}

\item{...}{Additional parameters (\code{CNV.load} generic, currently not used).}
}
\value{
\code{dataframe} object.
}
\description{
Load combined signal intensities from .idat-Files generated with the Illumina Mouse array. In the next step, use the resulting matrix for \code{CNV.load}.
}
\details{
This method loads the unmethylated and methylated signal intensity for each probe and sums them up. It is designed to be used for the Illumina Mouse arrays. Subsequently, the resulting matrix should be used for \code{CNV.load}
}
\author{
Bjarne Daenekas \email{conumee@hovestadt.bio}
}
