% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{CNV.merge_bins}
\alias{CNV.merge_bins}
\title{CNV.merge_bins}
\usage{
CNV.merge_bins(
  genome.anno,
  genome.tile,
  bin_minprobes = 15,
  genome.probes,
  bin_maxsize = 5e+06,
  verbose = FALSE
)
}
\arguments{
\item{genome.anno}{foo}

\item{genome.tile}{foo}

\item{bin_minprobes}{foo}

\item{genome.probes}{foo}

\item{bin_maxsize}{foo}

\item{verbose}{foo}
}
\value{
\code{GRanges} object.
}
\description{
Merge bins containing less than the defined number probes with neighboring bin containing fewer probes.
}
