% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{CNV.plotly}
\alias{CNV.plotly}
\title{CNV.plotly}
\usage{
CNV.plotly(xx, sample_n = 1)
}
\arguments{
\item{x}{an CNV analysis object generated with \code{\link[conumee]{CNV.segment}} or \code{\link[mnp.v11b4]{MNPcnv}}}
}
\description{
\code{MNPcnvggplotly} plot an interactive copy number profile
}
