% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{CNV.qqplot}
\alias{CNV.qqplot}
\alias{CNV.qqplot,CNV.analysis-method}
\title{CNV.qqplot}
\usage{
CNV.qqplot(object, ...)

\S4method{CNV.qqplot}{CNV.analysis}(
  object,
  sample = as.character(),
  gene = as.character(),
  conf = 0.99,
  minoverlap = 1L,
  set_par = TRUE,
  ...
)
}
\arguments{
\item{object}{\code{CNV.analysis} object.}

\item{...}{Additional parameters (\code{CNV.detailplot} generic, currently not used).}

\item{sample}{character. If \code{CNV.analysis} object contains multiple samples, please specify the sample name. You can check the sample names with \code{colnames(object@fit$ratio)}. If the object contains just one sample, this parameter can be left blank.}

\item{gene}{character. Please provide the gene symbol for the gene of interest. You can use this function for every gene that is part of the Cancer Gene Census (curated by the Sanger Institute).}

\item{conf}{numeric. This parameter affects the plotted confidence intervals. Which confidence level should be used? Default to \code{0.99}.}

\item{minoverlap}{integer. The function determines the bins that overlap with the genes of interest. Which minimum number of basepairs should be considered for an overlap? Defaul to \code{1L}.}

\item{set_par}{logical. Use recommended graphical parameters for \code{oma} and \code{mar}? Defaults to \code{TRUE}. Original parameters are restored afterwards.}
}
\value{
\code{NULL}.
}
\description{
Create a qqplot for the chromosome arm that contains the gene of interest. This plot should be used supportively to investigate diagnostically relevant high-level changes in the copy number.
}
\details{
This function creates a qqplot that illustrates if a gene of interest is part of the normal distribution of the bins' log2-ratios in the corresponding chromosome arm or if it should be considered as an outlier. The latter might indicate a high level change in copy number. To create the confidence intervals, the \code{BoutrosLab.plotting.general} package is used with \code{method = pointwise}.
}
\examples{

CNV.qqplot(x, gene = "EGFR", conf = 0.99, minoverlap = 10000L)

}
\author{
Bjarne Daenekas \email{conumee@hovestadt.bio}
}
