% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{CNV.state}
\alias{CNV.state}
\title{CNV.state}
\usage{
CNV.state(x, threshold = 0.1, deepthresh = 1)
}
\arguments{
\item{x}{a vector of segment CN values}

\item{threshold}{magnitude to call a loss or gain (0.1)}

\item{deepthresh}{magnitude for calling hiamp/deeploss (1)}
}
\value{
a vector of state calls using the above
}
\description{
helper function for converting +/- threshold to state (1/2/3) for CNVRanger
}
\examples{
Seg.CN <- c(A=-0.2, B=-0.1, C=-0.05, D=0, E=0.05, F=0.1, G=0.2)
CNV.state(Seg.CN, threshold=0.01)
CNV.state(Seg.CN, threshold=0.1)
CNV.state(Seg.CN, threshold=1)

}
