% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{CNV.summaryplot}
\alias{CNV.summaryplot}
\alias{CNV.summaryplot,CNV.analysis-method}
\title{CNV.summaryplot}
\usage{
CNV.summaryplot(object, ...)

\S4method{CNV.summaryplot}{CNV.analysis}(
  object,
  set_par = TRUE,
  main = NULL,
  output = "local",
  directory = getwd(),
  width = 12,
  height = 6,
  res = 720,
  threshold = 0.1,
  ...
)
}
\arguments{
\item{object}{\code{CNV.analysis} object.}

\item{...}{Additional parameters (\code{CNV.write} generic, currently not used).}

\item{set_par}{logical. Use recommended graphical parameters for \code{oma} and \code{mar}? Defaults to \code{TRUE}. Original parameters are restored afterwards.}

\item{main}{character. Specify the title of the plot. Defaults to \code{NULL}.}

\item{output}{character. Choose between \code{output} (to display it in the graphical output) or \code{pdf} and \code{png} to save it. Defaults to \code{local}.}

\item{directory}{character. Export directory for saving the files}

\item{width}{numeric. Width in inches of the saved files. Defaults to \code{12}.}

\item{height}{numeric. Height in inches of the saved files. Defaults to \code{8}}

\item{res}{numeric. Resolution of the saved .png files. Defaults to \code{720}}

\item{threshold}{numeric. Threshold for determining the copy number state. Defaults to \code{0.1}. See Details for details.}
}
\value{
\code{NULL}
}
\description{
Create a summaryplot that shows the CNVs in a all the query samples in the CNV.analysis object.
}
\details{
This function creates a plot that illustrates the changes in copy number states within the set of query samples that are stored in the \code{CNV.analysis} object. The y axis is showing the percentage of samples that are exhibiting a CNV at the genomic location shown on the x axis. The threshold for the log2-ratio to identify gains or losses is \code{0.1} by default.
}
\examples{
# prepare
library(minfiData)
data(MsetEx)
d <- CNV.load(MsetEx)
data(detail_regions)
anno <- CNV.create_anno(detail_regions = detail_regions)

# create/modify object
x <- CNV.segment(CNV.detail(CNV.bin(CNV.fit(query = d['GroupB_1'],
    ref = d[c('GroupA_1', 'GroupA_2', 'GroupA_3')], anno))))

# output plots
CNV.genomeplot(x)
CNV.genomeplot(x, chr = 'chr6')
CNV.detailplot(x, name = 'PTEN')
CNV.detailplot_wrap(x)
CNV.summaryplot(x, threshold = 0.15)
CNV.heatmap(x)

# output text files
CNV.write(x, what = 'segments')
CNV.write(x, what = 'detail')
CNV.write(x, what = 'bins')
CNV.write(x, what = 'probes')
}
\author{
Bjarne Daenekas \email{conumee@hovestadt.bio}
}
