% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{CNV.write}
\alias{CNV.write}
\alias{CNV.write,CNV.analysis-method}
\title{CNV.write}
\usage{
CNV.write(object, ...)

\S4method{CNV.write}{CNV.analysis}(object, file = NULL, what = "segments", threshold = 0.1)
}
\arguments{
\item{object}{\code{CNV.analysis} object.}

\item{...}{Additional parameters (\code{CNV.write} generic, currently not used).}

\item{file}{Path where output file should be written to. Defaults to \code{NULL}: No file is written, table is returned as data.frame object.}

\item{what}{character. This should be (an unambiguous abbreviation of) one of \code{'probes'}, \code{'bins'}, \code{'detail'}, \code{'segments'}, \code{gistic}, \code{overview} or \code{focal}. Defaults to \code{'segments'}.}

\item{threshold}{numeric. Threshold for determining the copy number state. Defaults to \code{0.1}. See Description for details.}
}
\value{
if parameter \code{file} is not supplied, the table is returned as a \code{data.frame} object.
}
\description{
Output CNV analysis results as table.
}
\details{
Function shows the output of the CNV analysis with conumee 2. To use the results as input for GISTIC choose \code{what = 'gistic'}. To assign the resulting segments to their copy number state and their size (focal, arm-level or whole chromosome) choose \code{what = 'overview'}. The threshold for the log2-ratio to identify gains or losses is \code{0.1} by default. To access the results from the Segmented Block Bootstrapping, use \code{what = focal}.
}
\examples{
# prepare
library(minfiData)
data(MsetEx)
d <- CNV.load(MsetEx)
data(detail_regions)
anno <- CNV.create_anno(detail_regions = detail_regions)

# create/modify object
x <- CNV.segment(CNV.detail(CNV.bin(CNV.fit(query = d['GroupB_1'],
    ref = d[c('GroupA_1', 'GroupA_2', 'GroupA_3')], anno))))

# output plots
CNV.genomeplot(x)
CNV.genomeplot(x, chr = 'chr6')
CNV.detailplot(x, name = 'PTEN')
CNV.detailplot_wrap(x)

# output text files
CNV.write(x, what = 'segments')
CNV.write(x, what = 'detail')
CNV.write(x, what = 'bins')
CNV.write(x, what = 'probes')
CNV.write(x, what = 'gistic')
CNV.write(x, what = 'overview')
CNV.write(x, what = 'focal')
}
\author{
Bjarne Daenekas, Volker Hovestadt \email{conumee@hovestadt.bio}
}
