% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{CNV.writebins}
\alias{CNV.writebins}
\title{CNV.writebins}
\usage{
CNV.writebins(object, file = NULL, threshold = 0.1)
}
\arguments{
\item{object}{\code{CNV.analysis} object.}

\item{file}{Path where output file should be written to. Defaults to \code{NULL}: No file is written, table is returned as data.frame object.}

\item{threshold}{numeric. This parameter is used internally for creating summaryplots. It should not be changed. If you intend to change the threshold for summaryplots, please do so within the \code{CNV.summaryplot} function.}
}
\value{
if parameter \code{file} is not supplied, the table is returned as a \code{data.frame} object.
}
\description{
Output bin-level CNV analysis results.
}
\examples{
library(minfiData)
data(MsetEx)
d <- CNV.load(MsetEx)
data(detail_regions)
anno <- CNV.create_anno(detail_regions = detail_regions)

x <- CNV.segment(CNV.detail(CNV.bin(CNV.fit(query = d['GroupB_1'],
    ref = d[c('GroupA_1', 'GroupA_2', 'GroupA_3')], anno))))

# output a data.frame
CNV.writebins(x)

# output text file
tmp <- tempfile(fileext=".igv")
CNV.writeprobes(x, file=tmp)

}
\author{
Tim Triche \email{trichelab@gmail.com}
}
