% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{CNV.writeoutput}
\alias{CNV.writeoutput}
\title{CNV.writeoutput}
\usage{
CNV.writeoutput(x, file = NULL)
}
\arguments{
\item{x}{results of CNV.write* methods (internally)}

\item{file}{path to file output. if NULL, return a data.frame.}
}
\value{
if \code{file} is NULL, return x as a \code{data.frame}.
}
\description{
helper function for CNV.write* methods.
}
\examples{
library(minfiData)
data(MsetEx)
d <- CNV.load(MsetEx)
data(detail_regions)
anno <- CNV.create_anno(detail_regions = detail_regions)

x <- CNV.segment(CNV.detail(CNV.bin(CNV.fit(query = d['GroupB_1'],
    ref = d[c('GroupA_1', 'GroupA_2', 'GroupA_3')], anno))))

# output a data.frame
out <- CNV.writebins(x)

# now output that to a file
tmp <- tempfile(fileext=".igv")
CNV.writeoutput(out, file=tmp)

}
\author{
Tim Triche \email{trichelab@gmail.com}
}
