% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{CNV.focal}
\alias{CNV.focal}
\alias{CNV.focal,CNV.analysis-method}
\title{CNV.focal}
\usage{
CNV.focal(object, ...)

\S4method{CNV.focal}{CNV.analysis}(object, conf = 0.99, minoverlap = 1000L, ...)
}
\arguments{
\item{object}{\code{CNV.analysis} object.}

\item{...}{Additional parameters (\code{CNV.detailplot} generic, currently not used).}

\item{conf}{numeric. This parameter affects the plotted confidence intervals. Which confidence level should be used? Default to \code{0.99}.}

\item{minoverlap}{integer. The function determines the bins that overlap with the genes of interest. Which minimum number of basepairs should be considered for an overlap? Defaul to \code{1000L}.}
}
\value{
A \code{CNV.analysis} object with significantly altered bins and genes from the Cancer Gene Census (curated by the Sanger Institute).
}
\description{
This optional function provides filtering for diagnostically relevant CNVs (high level amplification or homozygous deletion).
}
\details{
This function should facilitate the detection of diagnostically relevant CNVs that affect single genes. Therefore, a qqplot illustrating the bins' log2-ratios is created for each chromosome arm. In the first step, bins that lie outside the confidence interval are identified and sorted based on their residuals to the confidence curves.
Secondly, these bins are overlapped with the Cancer Gene Census to identify diagnostically relevant genes. The resulting bins and genes are sorted in regards to their significance.
}
\examples{

x <- CNV.focal(object, conf = 0.99, minoverlap = 10000L)
x@detail$cancer_genes
x@detail$significant_bins

}
\author{
Bjarne Daenekas \email{conumee@hovestadt.bio}
}
