% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{CNV.plotly}
\alias{CNV.plotly}
\title{CNV.plotly}
\usage{
CNV.plotly(x, sample = colnames(x@fit$coef)[1])
}
\arguments{
\item{x}{A \code{CNVanalysis} object after \code{CNV.segment} is performed.}

\item{sample}{character. Name of the single sample that should be plotted. Default to first sample in the set of query samples. Check sample names with \code{colnames(object@fit$ratio)}}
}
\description{
\code{CNV.plotly} plots an interactive copy number profile.
}
